asdfsdf

+++ client.c
@@ -6,11 +6,13 @@
 /*   By: yonshin <yonshin@student.42.fr>            +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2022/11/16 09:07:44 by yonshin           #+#    #+#             */
-/*   Updated: 2022/11/22 15:12:51 by yonshin          ###   ########.fr       */
+/*   Updated: 2022/11/22 17:16:13 by yonshin          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minitalk.h"
+
+static char *message;
 
 static int	get_pid(int argc, char *argv[])
 {
@@ -35,36 +37,46 @@ static char *get_message(int argc, char *argv[])
 	return (argv[2]);
 }
 
-void	signal_test1(int sig)
-{
-	ft_putstr_fd("sigusr1: ", 0);
-	ft_putnbr_fd(sig, 0);
-	ft_putchar_fd('\n', 0);
-}
-
 void	signal_test2(int sig, siginfo_t *info, void *old)
 {
-	signal_test1(sig);
-	ft_putnbr_fd(info->si_pid, 0);
-	ft_putchar_fd('\n', 0);
+	static t_sig	send;
+	int				res;
+
+	sig = 0;
 	old = 0;
-	kill(info->si_pid, SIGUSR1);
+	if (send.i == 0)
+		send.c = *message;
+	if (++send.i == END_OF_BYTE)
+	{
+		message++;
+		send.i = 0;
+	}
+	res = SIGUSR1 + (send.c & 0b10000000) / 0b10000000;
+	send.c = send.c << 1;
+	kill(info->si_pid, res);
+}
+
+void	signal_exit(int sig)
+{
+	if (sig == SIGUSR2)
+		exit(0);
 }
 
 int	main(int argc, char *argv[])
 {
 	const int			pid = get_pid(argc, argv);
-	const char			*message = get_message(argc, argv);
 	struct sigaction	usrsig;
 
+	message = get_message(argc, argv);
 	usrsig.sa_sigaction = signal_test2;
 	sigemptyset(&usrsig.sa_mask);
 	usrsig.sa_flags = 0;
 	usrsig.sa_flags = SA_NODEFER | SA_SIGINFO;
 	if (sigaction(SIGUSR1, &usrsig, 0) == -1)
 		exit(1);
+	if (signal(SIGUSR2, signal_exit) == SIG_ERR)
+		exit(1);
 	kill(pid, SIGUSR1);
-	message++;
 	while (1) ;
 	return (0);
 }
